requirements.txt
# django
django~=3.2.12
django_extensions

djangorestframework~=3.13.1
django-allauth~=0.51.0
django-cors-headers~=3.13.0
django-rest-auth~=0.9.5

# deploy toolkit
Fabric~=2.7.0

# Debug toolbar
django-debug-toolbar

# @phpdude libs
django-macros-url
django-template-names

# cache
python-memcached
django-clear-cache

# Database drivers. Uncomment anyone is used.
#MySQL-python
#mysqlclient
psycopg2

# Project custom apps
django-split-settings
Pillow
django-mathfilters
gunicorn==20.1.0
whitenoise==6.1.0
shortuuid==1.0.9